api/requirements.txt
asgiref==3.7.2
async-timeout==4.0.2
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.1
djangorestframework==3.14.0
exceptiongroup==1.1.1
fakeredis==2.15.0
idna==3.4
iniconfig==2.0.0
packaging==23.1
pluggy==1.2.0
pytest==7.4.0
pytest-django==4.5.2
pytz==2023.3
PyYAML==6.0
redis==4.6.0
requests==2.31.0
responses==0.23.1
sortedcontainers==2.4.0
sqlparse==0.4.4
tomli==2.0.1
types-PyYAML==6.0.12.10
typing_extensions==4.6.3
urllib3==2.0.3
